+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hipmi",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "private": true,
   "prisma": {
     "seed": "npx tsx prisma/seed.ts"
