+++ src/app/lib/global_state.ts
@@ -1,7 +1,9 @@
 import { atom } from "jotai";
+import { atomWithStorage } from "jotai/utils";
 
 export type IRealtimeData = {
-  status?: "Publish" | "Review" | "Draft" | "Reject";
+  status?: "Publish" | "Review" | "Draft" | "Reject" | "Peserta Event";
+  
   appId: string;
   userId: string;
   pesan: string;
@@ -14,14 +16,18 @@ export type IRealtimeData = {
     | "INVESTASI"
     | "COLLABORATION"
     | "FORUM";
-  userRole?: "USER" | "ADMIN";
 };
 
 export const gs_realtimeData = atom<IRealtimeData | null>(null);
 export const gs_admin_ntf = atom<number>(0);
 export const gs_user_ntf = atom<number>(0);
+export const gs_count_ntf = atom<number>(0);
+
 
 // job
+export const gs_adminJobTriggerReview = atom<boolean>(false);
+export const gs_jobTiggerBeranda = atom<boolean>(false);
 
-export const gs_job_trigger = atom<boolean>(false)
-
+// event
+export const gs_adminEventTriggerReview = atom<boolean>(false);
+export const gs_eventTriggerBeranda = atom<boolean>(false);
