+++ src/app_modules/admin/event/_view/index.ts
@@ -1,3 +1,5 @@
 import { AdminEvent_ViewDetailPeserta } from "./view_detail_peserta";
+import AdminEvent_ComponentTableReview from "./view_table_review";
 
-export { AdminEvent_ViewDetailPeserta  };
+export { AdminEvent_ViewDetailPeserta };
+export { AdminEvent_ComponentTableReview };
