+++ src/app_modules/admin/job/_view/index.ts
@@ -1,0 +1,3 @@
+import AdminJob_ViewTavleReview from "./view_table_review";
+
+export { AdminJob_ViewTavleReview };
