+++ src/app_modules/admin/notifikasi/fun/get/get_notifikasi_by_user_id.ts
@@ -3,10 +3,16 @@
 import prisma from "@/app/lib/prisma";
 import { funGetUserIdByToken } from "@/app_modules/_global/fun/get";
 
-export default async function adminNotifikasi_getByUserId() {
+export default async function adminNotifikasi_getByUserId({page}: {page: number}) {
   const userLoginId = await funGetUserIdByToken();
 
+
+  const takeData = 10;
+  const skipData = page * takeData - takeData;
+
   const data = await prisma.notifikasi.findMany({
+    take: takeData,
+    skip: skipData,
     orderBy: [
       {
         isRead: "asc",
