+++ src/app_modules/admin/notifikasi/fun/get/index.ts
@@ -1,0 +1,3 @@
+import { admin_funEventCheckStatus } from "./fun_event_check_status";
+
+export { admin_funEventCheckStatus };
