+++ src/app_modules/event/component/index.ts
@@ -1,3 +1,5 @@
+import Event_ComponentCreateButton from "./button/button_create_event";
 import { Event_ComponentSkeletonDetailData } from "./skeleton/comp_skeleton_detail_data";
 
 export { Event_ComponentSkeletonDetailData };
+export { Event_ComponentCreateButton };
