+++ src/app_modules/notifikasi/_ui/ui_event_notifikasi.tsx
@@ -11,6 +11,7 @@ import { ComponentNotifiaksi_CardView } from "../component";
 import notifikasi_getByUserId from "../fun/get/get_notifiaksi_by_id";
 import { gs_notifikasi_kategori_app } from "../lib";
 import { MODEL_NOTIFIKASI } from "../model/interface";
+import { useShallowEffect } from "@mantine/hooks";
 
 export default function Notifikasi_UiEvent({
   listNotifikasi,
@@ -20,6 +21,18 @@ export default function Notifikasi_UiEvent({
   const [data, setData] = useState<MODEL_NOTIFIKASI[]>(listNotifikasi);
   const [activePage, setActivePage] = useState(1);
   const [categoryPage, setCategoryPage] = useAtom(gs_notifikasi_kategori_app);
+
+  useShallowEffect(() => {
+    onLoadData(setData);
+  }, [setData]);
+
+  async function onLoadData(setData: any) {
+    const listNotifikasi = await notifikasi_getByUserId({
+      page: 1,
+      kategoriApp: "Event",
+    });
+    setData(listNotifikasi);
+  }
 
   return (
     <>
