+++ src/app_modules/notifikasi/_view/index.ts
@@ -1,3 +1,3 @@
-import Notifikasi_ViewNewMain from "./view_new_notifikasi";
+import Notifikasi_ViewNewMain from "./view_card_job";
 
 export { Notifikasi_ViewNewMain };
