+++ src/app_modules/notifikasi/_view/view_card_job.tsx
@@ -1,0 +1,6 @@
+export default function Notifikasi_ViewCardJob(){
+    return<>
+    
+    
+    </>
+}
