--- src/app_modules/notifikasi/_view/view_new_notifikasi.tsx
@@ -1,6 +1,0 @@
-export default function Notifikasi_ViewNewMain(){
-    return<>
-    
-    
-    </>
-}
