+++ src/app_modules/notifikasi/fun/count/fun_count_by_id.ts
@@ -1,7 +1,9 @@
 "use server";
 
 import prisma from "@/app/lib/prisma";
+import { RouterHome } from "@/app/lib/router_hipmi/router_home";
 import { funGetUserIdByToken } from "@/app_modules/_global/fun/get";
+import { revalidatePath } from "next/cache";
 
 export default async function notifikasi_countUserNotifikasi() {
   const userLoginId = await funGetUserIdByToken();
@@ -14,5 +16,6 @@ export default async function notifikasi_countUserNotifikasi() {
     },
   });
 
+  revalidatePath(RouterHome.main_home);
   return count.length;
 }
